+++ CHANGE-LOG.md
@@ -4,7 +4,7 @@
 - More documentation (**FEATURE**)
 - Include git short hash in build (**FEATURE**)
 - Use IDictionary instead of Dictionary for serialization (**FEATURE**)
-
+- Support single value deserialization from dictionary (**FEATURE**)
 
 ## Version 0.9.0 (released 2019-02-03)
 - Initial Nuget release